+++ tomcat/src/main/java/http/header/ContentType.java
@@ -1,4 +1,4 @@
-package http;
+package http.header;
 
 import java.util.Arrays;
