+++ tomcat/src/main/java/http/header/HttpHeaders.java
@@ -1,4 +1,4 @@
-package http;
+package http.header;
 
 import java.util.ArrayList;
 import java.util.HashMap;
