+++ tomcat/src/main/java/http/request/HttpRequestStartLine.java
@@ -1,4 +1,4 @@
-package http;
+package http.request;
 
 import http.exception.InvalidHttpRequestFormatException;
 import java.net.URI;
