+++ tomcat/src/main/java/nextstep/jwp/presentation/LoginController.java
@@ -1,9 +1,9 @@
 package nextstep.jwp.presentation;
 
-import http.HttpHeaders;
-import http.HttpRequest;
-import http.HttpResponse;
-import http.HttpStatus;
+import http.header.HttpHeaders;
+import http.request.HttpRequest;
+import http.response.HttpResponse;
+import http.response.HttpStatus;
 import http.session.Session;
 import http.session.SessionManager;
 import java.util.UUID;
