+++ tomcat/src/main/java/org/apache/catalina/handler/RequestHandler.java
@@ -1,7 +1,7 @@
 package org.apache.catalina.handler;
 
-import http.HttpRequest;
-import http.HttpResponse;
+import http.request.HttpRequest;
+import http.response.HttpResponse;
 
 public interface RequestHandler {
