+++ tomcat/src/test/java/http/HttpRequestStartLineTest.java
@@ -5,6 +5,8 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import http.exception.InvalidHttpRequestFormatException;
+import http.request.HttpMethod;
+import http.request.HttpRequestStartLine;
 import java.net.URI;
 import org.junit.jupiter.api.Test;
