+++ tomcat/src/test/java/http/header/ContentTypeTest.java
@@ -1,4 +1,4 @@
-package http;
+package http.header;
 
 import static org.assertj.core.api.Assertions.assertThat;
