+++ tomcat/src/test/java/http/repsonse/HttpResponseTest.java
@@ -1,8 +1,11 @@
-package http;
+package http.repsonse;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
+import http.header.HttpHeaders;
+import http.response.HttpResponse;
+import http.response.HttpStatus;
 import org.junit.jupiter.api.Test;
 
 class HttpResponseTest {
