+++ tomcat/src/test/java/http/request/QueryParamsTest.java
@@ -1,4 +1,4 @@
-package http;
+package http.request;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
