+++ tomcat/src/test/java/nextstep/jwp/presentation/RegisterControllerTest.java
@@ -3,8 +3,8 @@ package nextstep.jwp.presentation;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import http.HttpRequest;
-import http.HttpResponse;
+import http.request.HttpRequest;
+import http.response.HttpResponse;
 import java.io.BufferedReader;
 import java.io.ByteArrayInputStream;
 import java.io.File;
